--- wls-gui-wahllokalsystem/src/types/stimmabgabermerke/EingenommenerWahlschein.ts
@@ -1,6 +1,0 @@
-import type { EingenommenerWahlscheinStimmzettelartEnum } from "@/types/stimmabgabermerke/EingenommenerWahlscheinStimmzettelartEnum.ts";
-
-export interface EingenommenerWahlschein {
-  anzahl: number;
-  stimmzettelart: EingenommenerWahlscheinStimmzettelartEnum;
-}
