+++ wls-gui-wahllokalsystem/src/types/stimmabgabevermerke/EingenommenerWahlschein.ts
@@ -1,0 +1,6 @@
+import type { EingenommenerWahlscheinStimmzettelartEnum } from "@/types/stimmabgabevermerke/EingenommenerWahlscheinStimmzettelartEnum.ts";
+
+export interface EingenommenerWahlschein {
+  anzahl: number;
+  stimmzettelart: EingenommenerWahlscheinStimmzettelartEnum;
+}
