+++ wls-gui-wahllokalsystem/src/types/stimmabgabevermerke/EingenommenerWahlscheinStimmzettelartEnum.ts
