+++ wls-gui-wahllokalsystem/src/types/stimmabgabevermerke/Wahldaten.ts
@@ -1,4 +1,4 @@
-import type { EingenommenerWahlscheinStimmzettelartEnum } from "@/types/stimmabgabermerke/EingenommenerWahlscheinStimmzettelartEnum.ts";
+import type { EingenommenerWahlscheinStimmzettelartEnum } from "@/types/stimmabgabevermerke/EingenommenerWahlscheinStimmzettelartEnum.ts";
 
 export interface Wahldaten {
   wahlbezirkID: string;
